+++ azuredevops/internal/provider/framework_provider.go
@@ -20,6 +20,7 @@ import (
 	"github.com/parsoFish/terraform-provider-betterado/azuredevops/internal/service/memberentitlementmanagement"
 	"github.com/parsoFish/terraform-provider-betterado/azuredevops/internal/service/permissions"
 	"github.com/parsoFish/terraform-provider-betterado/azuredevops/internal/service/release"
+	"github.com/parsoFish/terraform-provider-betterado/azuredevops/internal/service/servicehook"
 	"github.com/parsoFish/terraform-provider-betterado/azuredevops/internal/service/taskagent"
 )
 
@@ -225,6 +226,8 @@ func (p *BetteradoFrameworkProvider) Resources(_ context.Context) []func() resou
 		git.NewGitRepositoryResource,
 		git.NewGitRepositoryBranchResource,
 		git.NewGitRepositoryFileResource,
+		servicehook.NewServicehookStorageQueuePipelinesResource,
+		servicehook.NewServicehookWebhookTfsResource,
 	}
 }